aashray/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
boto3==1.16.55
botocore==1.19.63
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.5
django-import-export==2.5.0
django-storages==1.11.1
et-xmlfile==1.0.1
filelock==3.0.12
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.4
virtualenv==20.4.3
xlrd==2.0.1
xlwt==1.3.0
